LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-self-allow-expected.txt

CONSOLE MESSAGE: line 60: IFrame load event fired: the IFrame's location is 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy=%27self%27'.
A 'frame-ancestors' CSP directive with a value 'self' should allow rendering.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The IFrame should not have been blocked. It wasn't.
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
This is an IFrame sending a Content Security Policy header containing "frame-ancestors 'self'".
